// cypress/e2e/interfaces/information.ts
interface Information {
  name: string;
  lastName: string;
  email: string;
  gender: string;
  mobileNumber: string;
  subjects: string;
  hobbies: string[];
  currentAddress: string;
  state: string;
  city: string;
  successfulMessage: string;
  studentNameLabel: string;
  studentEmailLabel: string;
  genderLabel: string;
  mobileLabel: string;
  subjectsLabel: string;
  hobbiesLabel: string;
  addressLabel: string;
  stateAndCityLabel: string;
}
export type {Information}